append test

+++ DroidKit/Extension/UIColor+Extensions.swift
@@ -17,9 +17,9 @@ extension UIColor {
     /// example
     /// UIColor(hex: 0xAABBCC)
     convenience init(hex: Int, alpha: Double = 1.0) {
-        let r = CGFloat((hex & 0xFF0000) >> 16) / 255.0
-        let g = CGFloat((hex & 0x00FF00) >> 8) / 255.0
-        let b = CGFloat(hex & 0x0000FF) / 255.0
+        let r = CGFloat((hex & 0xFF0000) >> 16) / CGFloat(255.0)
+        let g = CGFloat((hex & 0x00FF00) >> 8) / CGFloat(255.0)
+        let b = CGFloat(hex & 0x0000FF) / CGFloat(255.0)
         self.init(red: r, green: g, blue: b, alpha: CGFloat(alpha))
     }
